refactor(routes): 페이지 라우트를 src/app에서 app으로 이동
 Leave 페이지를 app/leave로 이동 (LeaveReasonDropdown, layout, page)
 Onboarding 페이지를 app/onboarding로 이동 (KakaoLoginButton, Logo, page)
 Smarttalk 페이지를 app/smarttalk로 이동 (layout, page)
 Welcome 페이지를 app/welcome로 이동 (RandomNickname, page)
 Import 경로를 path-alias를 사용한 절대 경로로 변경
 Smarttalk UI 컴포넌트 추가 (SmarttalkPage)
 Next.js App Router 구조에 맞게 라우트 정리
