+++ apps/web/app/leave/LeaveReasonDropdown.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/utils/style';
+import { cn } from '@/shared/utils/style';
 import { LEAVE_REASON } from '@/shared/config/leaveReason';
 import { useEffect, useRef, useState } from 'react';
