+++ apps/web/app/onboarding/Logo.tsx
