+++ apps/web/app/welcome/RandomNickname.tsx
@@ -2,14 +2,14 @@
 
 import type { TokenPayload } from '@moneed/auth';
 import { decodeJwt } from 'jose';
-import { getCookie } from '@/shared/utils/cookie.browser';
+import { getToken } from '@/shared/utils/cookie.browser';
 import { useEffect, useState } from 'react';
 
 export default function RandomNickname() {
     const [nickname, setNickname] = useState<string>('');
 
     useEffect(() => {
-        const accessToken = getCookie('access_token');
+        const accessToken = getToken('access_token');
         if (!accessToken) {
             return;
         }
